feat: initial refactor of Import (WIP)
