+++ apps/antalmanac/src/components/Header/Header.tsx
@@ -7,7 +7,7 @@ import { useState, type MouseEventHandler } from 'react';
 
 import AboutPage from './AboutPage';
 import Feedback from './Feedback';
-import ImportStudyList from './ImportStudyList';
+import Import from './Import';
 import LoadSaveScheduleFunctionality from './LoadSaveFunctionality';
 import SettingsMenu from './SettingsMenu';
 import Export from './Export';
@@ -42,7 +42,7 @@ interface CustomAppBarProps {
 }
 
 const components = [
-    <ImportStudyList key="studylist" />,
+    <Import key="studylist" />,
     <Export key="export" />,
     <Feedback key="feedback" />,
     <AboutPage key="about" />,
